audio: set HTTP stream timeout after connect
In commit 67ff0351b741 ("audio: increase HTTP stream timeout after
send") we increased the timeout for the HTTP stream to WIS STT after
send to 10s, to allow people with slow WIS to still use Willow. For some
reason this worked in the 0.3 releases, but it seems this doesn't work
anymore after all the bumps of Espressif components in the 0.4 releases.

The commit was wrong in the first place, as we should increase the
timeout after initial connect, but before we finish sending, as setting
it after we finished sending should result in the connection timing out
after 2s while sending.

Set the timeout on the HTTP_STREAM_ON_REQUEST event to fix this.

Fixes: 67ff0351b741 ("audio: increase HTTP stream timeout after send")